hitl-module/services/epub-conversion/pyproject.toml
[tool.poetry]
name = "epub-conversion"
version = "0.1.0"
description = "EPUB conversion service scaffold"
authors = ["Engineering <engineering@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.32.0" }
pydantic = "^2.10.0"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
alembic = "^1.14.0"
greenlet = "^3.1.1"
redis = ">=5.0"
boto3 = ">=1.26"
pdfminer-six = ">=20221105"
pymupdf = ">=1.23"
ebooklib = ">=0.18"
pillow = ">=10.0"
python-multipart = ">=0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24"
python-docx = "^1.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
